Finished running 4 diff simulations: 2 with 30 degree rotations, with and without integration of training examples across 2nd layer of network.  Signifiicant difference between 30 degree and 0 degrees as to training and test error.  This is understandable: 30 degrees puts most of the differentiating information into one of the 2 input variables.
Separately, and most important, is the fact that, for both rotations, the test error was significantly better for Proj4.

Have made some minor mods for operation with Docker and AWS execution.  Continuing with experiments using GitHub as code base for simulations on docker.

Now adding simplest logger (StringIO and FileIO), that will feed into Redis, and then possibly mongo.  KEEP IT SIMPLE>>>

getting urls and $currenthost doing the right thinigs...