Finished running 4 diff simulations: 2 with 30 degree rotations, with and without integration of training examples across 2nd layer of network.  Signifiicant difference between 30 degree and 0 degrees as to training and test error.  This is understandable: 30 degrees puts most of the differentiating information into one of the 2 input variables.
Separately, and most important, is the fact that, for both rotations, the test error was significantly better for Proj4.

Have made some minor mods for operation with Docker and AWS execution.  Continuing with experiments using GitHub as code base for simulations on docker.

Now adding simplest logger (StringIO and FileIO), that will feed into Redis, and then possibly mongo.  KEEP IT SIMPLE>>>

getting urls and $currenthost doing the right thinigs... AGAIN and saving again  REMOTE redis storage working just fine now.

Both Proj3 and Proj4 working with Redis and environment variables.

*** All std Projects working:  BasicBPDemo, Proj3, Proj4, and Proj3comparedTo4  I'm about to start designing/coding the cat's meow.

Just did a significant amount of reshuffling code across files.  Some renaming of .rb files, also.  All bins work...

Woops, I need to reverse about an hours worth of editing in the Learning Strategy file.!!

All std Projects working again:  I forgot the "self." in "self.netInput"!

spent a little time with docker and redis -- to get persistence between rebuilds and reruns.

Now testing the continuous use of "alwaysOn" server for Redis storage of simuluation results........

Ok, the server and redis stuff is done for now.  Now I've been adding code for Symmetrical Models for Self-Org and Supervised.  Also considering stochastic switching of neuron outputs according to the probability that the example is from one of the two lobe (Gauss Models).  In other words, if an example is unlikely "from" the lobe models, the example should be "ignored by that neuron" -- or, the neuron should indicate that it "has NO OPINION"!!

Made significant progress on mixture model for the distribuion of a neuron's netInput or outputs.  netInputs or outputs depends on... the learning strategy for effectively and adaptively incorporate sigmoid (and other) nonlinearities in the network's processing.